TRANSACTION : - security - refresh api token
PROFILE :
- crud

+++ src/Entity/Recipe.php
@@ -123,7 +123,7 @@ class Recipe
     private $number_of_persons;
 
     /**
-     * @ORM\Column(type="float")
+     * @ORM\Column(type="integer")
      * @Groups({"recipe:list", "recipe:item"})
      */
     private $difficulty;
@@ -274,12 +274,12 @@ class Recipe
         return $this;
     }
 
-    public function getDifficulty(): ?float
+    public function getDifficulty(): ?int
     {
         return $this->difficulty;
     }
 
-    public function setDifficulty(float $difficulty): self
+    public function setDifficulty(int $difficulty): self
     {
         $this->difficulty = $difficulty;
